Removing Old Code.
More changes with the core working. The color edit will take a lot of time as i want to be able to print the "NAME" of the color and cycle through pre determined colors (or patterns of colors) then save that on a per pad basis. So the Snare can be Rainbow and the tom and be solid GREEN. More research needed into Color Pallets or arrays of colors/patterns.